Calculator.add() signature changed, it now operates on doubles

+++ calculator/src/main/java/pl/jug/szczecin/workshop/Calculator.java
@@ -2,7 +2,7 @@ package pl.jug.szczecin.workshop;
 
 public class Calculator {
 
-    public int add(int a, int b) {
+    public double add(double a, double b) {
         return a + b;
     }
